requirements.txt
# Data processing
pandas==2.1.0
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.0
tensorflow==2.13.0
xgboost==2.0.0

# Data visualization
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dotenv==1.0.0
mysql-connector-python==8.1.0
joblib==1.3.2

# API
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6 